Rename the imagevolume e2e suite to combinedvolumes
